+++ x/eibc/keeper/hooks.go
@@ -61,10 +61,7 @@ func (d delayedAckHooks) AfterPacketDeleted(ctx sdk.Context, rollappPacket *comm
 	// Get the demand order from the packet key. The initial demand order was built when
 	// the packet was created, hence with PENDING status.
 	rollappPacket.Status = commontypes.Status_PENDING
-	packetKey, err := commontypes.RollappPacketKey(rollappPacket)
-	if err != nil {
-		return err
-	}
+	packetKey := commontypes.RollappPacketKey(rollappPacket)
 	demandOrderID := types.BuildDemandIDFromPacketKey(string(packetKey))
 	demandOrder, err := d.GetDemandOrder(ctx, commontypes.Status_FINALIZED, demandOrderID)
 	if err != nil {