+++ x/eibc/keeper/hooks_test.go
@@ -16,8 +16,8 @@ func (suite *KeeperTestSuite) TestAfterRollappPacketUpdated() {
 	// Create new demand order
 	demandOrderFulfillerAddr := apptesting.AddTestAddrs(suite.App, suite.Ctx, 1, math.NewInt(1000))[0].String()
 	demandOrder, err := types.NewDemandOrder(*rollappPacket, "100", "50", sdk.DefaultBondDenom, demandOrderFulfillerAddr)
-	suite.Require().Equal(demandOrder.TrackingPacketStatus, commontypes.Status_PENDING)
 	suite.Require().NoError(err)
+	suite.Require().Equal(demandOrder.TrackingPacketStatus, commontypes.Status_PENDING)
 	err = suite.App.EIBCKeeper.SetDemandOrder(suite.Ctx, demandOrder)
 	suite.Require().NoError(err)
 	// Update rollapp packet status to finalized
@@ -27,7 +27,7 @@ func (suite *KeeperTestSuite) TestAfterRollappPacketUpdated() {
 	updatedDemandOrder, err := suite.App.EIBCKeeper.GetDemandOrder(suite.Ctx, commontypes.Status_FINALIZED, demandOrder.Id)
 	suite.Require().NoError(err)
 	suite.Require().Equal(updatedDemandOrder.TrackingPacketStatus, commontypes.Status_FINALIZED)
-	rollappPacketKey, err := commontypes.RollappPacketKey(&updatedRollappPacket)
+	rollappPacketKey := commontypes.RollappPacketKey(&updatedRollappPacket)
 	suite.Require().NoError(err)
 	suite.Require().Equal(updatedDemandOrder.TrackingPacketKey, string(rollappPacketKey))
 }